完成一部分的单角色到多角色的重构

+++ src/main/java/estate/dao/PropertyDao.java
@@ -42,7 +42,7 @@ public interface PropertyDao
      * @param role
      * @return
      */
-    ArrayList<PropertyEntity> getPropertiesByPhoneRole(String phone,int role);
+    ArrayList<PropertyEntity> getPropertiesByPhoneRole(String phone,Byte role);
 
 
     /**
